refactor(bash): extract process-execution layer + split tests (dirge-4y4l stage 9a)
Begin decomposing the top god-file by structural irregularity:
src/agent/tools/bash.rs (2102 LOC). Convert the flat file into a
bash/ directory module and split out two cohesive concerns.

- bash/exec.rs (374 LOC) — process-execution layer: InterleavedOutput,
  PgKillGuard, run_with_timeout (bounded synchronous runs), and
  spawn_streaming_shell (detached background runs). Self-contained:
  depends only on ToolError, tokio Command/Duration, bg_shell, and libc;
  nothing from BashTool or the parsing helpers. run_with_timeout /
  spawn_streaming_shell are pub(super), re-imported into mod.rs.
- bash/tests.rs (1062 LOC) — the #[cfg(test)] #[cfg(unix)] module,
  moved verbatim. Adds explicit Command/Duration imports it previously
  inherited from mod.rs's top-level `use`.
- bash/mod.rs (684 LOC, was 2102) — BashTool struct/impls, the Tool
  impl, and the command-parsing / permission-check cluster (deferred to
  stage 9b).

Only BashTool is public externally (pub use bash::BashTool in tools/mod.rs)
so the file->directory move is transparent to callers. Behavior-preserving.
-D warnings clean default + windows-default; 2374 tests green.

Refs dirge-5lh5
